+++ src/components/MenuBars/MenuItem/MenuToggle.tsx
@@ -1,31 +1,33 @@
-import React, { useState } from 'react';
+import { useState } from 'react';
 import {ReactComponent as CloseIcon} from "assets/icon-close.svg";
+import classNames from 'classnames';
 import './MenuItem.scss';
 
 type MenuToggleProps = {
-    children: React.ReactNode;
     direction: 'left' | 'right';
-    index: number;
     onToggle?: (active: boolean) => void;
     toggleStartLabel: string;
     toggleStopLabel: string;
+    icon: React.FunctionComponent<React.SVGProps<SVGSVGElement>>;
 }
 
 function MenuToggle(props: MenuToggleProps) {
 
-    const [status, setStatus] = useState(false);
+    const [isActive, setStatus] = useState(false);
+
+    const Icon = props.icon;
 
     return (<button 
-                className={`menu-item menu-item${props.index} menu-item${status ? '--active' : '--disabled'} menu-item--${props.direction}`}
+                className={classNames('menu-item', {'menu-item__active': isActive, 'menu-item__disabled': !isActive}, `menu-item--${props.direction}`)}
                 onClick={_ => setStatus(prevValue => !prevValue)}
             >
         <div className='menu-item__tooltip'>
             <span className='tooltip__text'>
-                {status ? props.toggleStopLabel : props.toggleStartLabel}
+                {isActive ? props.toggleStopLabel : props.toggleStartLabel}
             </span>
         </div>
+        <Icon className='menu-item__icon menu-item__icon--start'/>
         <CloseIcon className='menu-item__icon menu-item__icon--end'/>
-        {props.children}
     </button>);
 }
