+++ src/constants/style.scss
@@ -20,6 +20,11 @@ $color-pink: #F20574;
 .accent-color__violet {--accent-color: #AB05F2;}
 .accent-color__pink {--accent-color: #F20574;}
 
+// menu colors
+$menu-color-blue--primary: #0057ff;
+$menu-color-blue--secondary: #253f71;
+$menu-color-pink--primary: #ff0069;
+$menu-color-pink--secondary: #7c123e;
 
 // board & column constants
 $board__side-panel-width: 64px;
