+++ src/index.scss
@@ -5,6 +5,7 @@ html, body, #root {
 }
 *, *:before, *:after {
   box-sizing: inherit;
+  font-family: 'Raleway', sans-serif;
 }
 
 body {
